Add cashscript/jest export and make TestExtensions work with jest and vitest
