+++ packages/cashscript/.npmignore
@@ -1,4 +1,5 @@
 *
 !dist/**
 !vitest/package.json
+!jest/package.json
 *.map
