+++ packages/cashscript/jest/package.json
