+++ packages/cashscript/tsconfig.build.json
@@ -2,11 +2,14 @@
   "extends": "../../tsconfig.build.json",
   "compilerOptions": {
     "outDir": "./dist",
-    "types": ["vitest/globals"]
+    "types": [
+      "vitest/globals"
+    ]
   },
   "include": [
     "src/**/*",
-    "vitest.d.ts",
   ],
-  "exclude": ["test/fixture"],
+  "exclude": [
+    "test/fixture"
+  ],
 }
